fix(starboard): issues with self_star and star_starboard, message deletion

+++ src/modules/features/starboard/messages/deleteStarredMessage.ts
@@ -12,14 +12,14 @@ export default async function deleteStarredMessage(
    const starred = server.starred_messages.find(
       (i) => i.starred_message_id == messageReaction.message.id && i.board == board.board_name,
    );
+   if (!starred) return;
    const starboard_channel: GuildBasedChannel | undefined = await messageReaction.message.guild.channels
       .fetch(board.channelID)
       .catch(() => undefined);
    if (!starboard_channel || !starboard_channel.isTextBased()) return;
-   const message = await starboard_channel.messages.fetch(starred.starboard_message_id).catch(() => undefined);
+   const message = await starboard_channel.messages.fetch(starred?.starboard_message_id).catch(() => undefined);
    try {
-      if (!message) return;
-      await message.delete();
+      if (message) await message.delete();
       server.starred_messages.splice(server.starred_messages.indexOf(starred), 1);
       await auxdibot.database.servers.update({
          where: { serverID: messageReaction.message.guild.id },
